Added open & talk verbs
Added open verb as well as some handling of updating room objects as they are revealed by open objects. (Did this for things like "pull" as well. There's a few scenarios of this...)

Also added "talk" option. This is primarily handled by the "chat" functions in the helper_functions.py to save space. Not sure this is the best way to handle those, so open for feedback here!!